requirements.txt
boto==2.46.1
Django==1.9
django-adminplus==0.5
django-cors-headers==2.0.2
django-ses==0.8.1
django-smtp-ssl==1.0
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
djoser==0.5.4
gunicorn==19.6.0
mysqlclient==1.3.9
PyJWT==1.4.2
PyMySQL==0.7.9
virtualenv==15.1.0
